LayoutTests/fast/forms/ios/user-scalable-does-not-scale-for-keyboard-focus-with-author-defined-scale.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<meta name="viewport" content="width=device-width, maximum-scale=1">
</head>

<body onload="runTest();">
<input type="text" id="textfield">
<p id="result">This test has to run in iOS WebKitTestRunner.</p>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that even though force user scalable = true, we won't scale if a text field gets focus and the page has maximum-scale=1");

    jsTestIsAsync = true;
    if (window.testRunner)
        testRunner.setIgnoresViewportScaleLimits(true);

    async function runTest()
    {
        if (testRunner.runUIScript) {
            await UIHelper.activateAndWaitForInputSessionAt(10, 10);
            await UIHelper.waitForZoomingOrScrollingToEnd();
            zoomScaleAfterFocusing = await UIHelper.zoomScale();
            shouldBe("zoomScaleAfterFocusing", "1");
            finishJSTest();
        }
    }
</script>

</body>
</html>
